chore: remove dead operation code
NODE-1962
