+++ lib/collection.js
@@ -27,13 +27,11 @@ const CommandCursor = require('./command_cursor');
 // Operations
 const checkForAtomicOperators = require('./operations/collection_ops').checkForAtomicOperators;
 const ensureIndex = require('./operations/collection_ops').ensureIndex;
-const findAndModify = require('./operations/collection_ops').findAndModify;
-const findAndRemove = require('./operations/collection_ops').findAndRemove;
 const group = require('./operations/collection_ops').group;
 const parallelCollectionScan = require('./operations/collection_ops').parallelCollectionScan;
-const removeDocuments = require('./operations/collection_ops').removeDocuments;
+const removeDocuments = require('./operations/common_functions').removeDocuments;
 const save = require('./operations/collection_ops').save;
-const updateDocuments = require('./operations/collection_ops').updateDocuments;
+const updateDocuments = require('./operations/common_functions').updateDocuments;
 
 const AggregateOperation = require('./operations/aggregate');
 const BulkWriteOperation = require('./operations/bulk_write');
@@ -48,6 +46,7 @@ const DropIndexOperation = require('./operations/drop_index');
 const DropIndexesOperation = require('./operations/drop_indexes');
 const EstimatedDocumentCountOperation = require('./operations/estimated_document_count');
 const FindOneOperation = require('./operations/find_one');
+const FindAndModifyOperation = require('./operations/find_and_modify');
 const FindOneAndDeleteOperation = require('./operations/find_one_and_delete');
 const FindOneAndReplaceOperation = require('./operations/find_one_and_replace');
 const FindOneAndUpdateOperation = require('./operations/find_one_and_update');
@@ -1658,14 +1657,11 @@ function _findAndModify(query, sort, doc, options, callback) {
   // Force read preference primary
   options.readPreference = ReadPreference.PRIMARY;
 
-  return executeLegacyOperation(this.s.topology, findAndModify, [
-    this,
-    query,
-    sort,
-    doc,
-    options,
+  return executeOperation(
+    this.s.topology,
+    new FindAndModifyOperation(this, query, sort, doc, options),
     callback
-  ]);
+  );
 }
 
 /**
@@ -1688,13 +1684,14 @@ Collection.prototype.findAndRemove = deprecate(function(query, sort, options, ca
   sort = args.length ? args.shift() || [] : [];
   options = args.length ? args.shift() || {} : {};
 
-  return executeLegacyOperation(this.s.topology, findAndRemove, [
-    this,
-    query,
-    sort,
-    options,
+  // Add the remove option
+  options.remove = true;
+
+  return executeOperation(
+    this.s.topology,
+    new FindAndModifyOperation(this, query, sort, null, options),
     callback
-  ]);
+  );
 }, 'collection.findAndRemove is deprecated. Use findOneAndDelete instead.');
 
 /**
