+++ lib/operations/collection_ops.js
@@ -1,8 +1,6 @@
 'use strict';
 
 const applyWriteConcern = require('../utils').applyWriteConcern;
-const applyRetryableWrites = require('../utils').applyRetryableWrites;
-const checkCollectionName = require('../utils').checkCollectionName;
 const Code = require('../core').BSON.Code;
 const createIndexDb = require('./db_ops').createIndex;
 const decorateWithCollation = require('../utils').decorateWithCollation;
@@ -10,8 +8,6 @@ const decorateWithReadConcern = require('../utils').decorateWithReadConcern;
 const ensureIndexDb = require('./db_ops').ensureIndex;
 const evaluate = require('./db_ops').evaluate;
 const executeCommand = require('./db_ops').executeCommand;
-const executeDbAdminCommand = require('./db_ops').executeDbAdminCommand;
-const formattedOrderClause = require('../utils').formattedOrderClause;
 const resolveReadPreference = require('../utils').resolveReadPreference;
 const handleCallback = require('../utils').handleCallback;
 const indexInformationDb = require('./db_ops').indexInformation;
@@ -19,14 +15,8 @@ const Long = require('../core').BSON.Long;
 const MongoError = require('../core').MongoError;
 const ReadPreference = require('../core').ReadPreference;
 const toError = require('../utils').toError;
-
-let collection;
-function loadCollection() {
-  if (!collection) {
-    collection = require('../collection');
-  }
-  return collection;
-}
+const insertDocuments = require('./common_functions').insertDocuments;
+const updateDocuments = require('./common_functions').updateDocuments;
 
 /**
  * Group function helper
@@ -61,98 +51,6 @@ function loadCollection() {
 // }.toString();
 const groupFunction =
   'function () {\nvar c = db[ns].find(condition);\nvar map = new Map();\nvar reduce_function = reduce;\n\nwhile (c.hasNext()) {\nvar obj = c.next();\nvar key = {};\n\nfor (var i = 0, len = keys.length; i < len; ++i) {\nvar k = keys[i];\nkey[k] = obj[k];\n}\n\nvar aggObj = map.get(key);\n\nif (aggObj == null) {\nvar newObj = Object.extend({}, key);\naggObj = Object.extend(newObj, initial);\nmap.put(key, aggObj);\n}\n\nreduce_function(obj, aggObj);\n}\n\nreturn { "result": map.values() };\n}';
-
-/**
- * Perform a bulkWrite operation. See Collection.prototype.bulkWrite for more information.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object[]} operations Bulk operations to perform.
- * @param {object} [options] Optional settings. See Collection.prototype.bulkWrite for a list of options.
- * @param {Collection~bulkWriteOpCallback} [callback] The command result callback
- */
-function bulkWrite(coll, operations, options, callback) {
-  // Add ignoreUndefined
-  if (coll.s.options.ignoreUndefined) {
-    options = Object.assign({}, options);
-    options.ignoreUndefined = coll.s.options.ignoreUndefined;
-  }
-
-  // Create the bulk operation
-  const bulk =
-    options.ordered === true || options.ordered == null
-      ? coll.initializeOrderedBulkOp(options)
-      : coll.initializeUnorderedBulkOp(options);
-
-  // Do we have a collation
-  let collation = false;
-
-  // for each op go through and add to the bulk
-  try {
-    for (let i = 0; i < operations.length; i++) {
-      // Get the operation type
-      const key = Object.keys(operations[i])[0];
-      // Check if we have a collation
-      if (operations[i][key].collation) {
-        collation = true;
-      }
-
-      // Pass to the raw bulk
-      bulk.raw(operations[i]);
-    }
-  } catch (err) {
-    return callback(err, null);
-  }
-
-  // Final options for retryable writes and write concern
-  let finalOptions = Object.assign({}, options);
-  finalOptions = applyRetryableWrites(finalOptions, coll.s.db);
-  finalOptions = applyWriteConcern(finalOptions, { db: coll.s.db, collection: coll }, options);
-
-  const writeCon = finalOptions.writeConcern ? finalOptions.writeConcern : {};
-  const capabilities = coll.s.topology.capabilities();
-
-  // Did the user pass in a collation, check if our write server supports it
-  if (collation && capabilities && !capabilities.commandsTakeCollation) {
-    return callback(new MongoError('server/primary/mongos does not support collation'));
-  }
-
-  // Execute the bulk
-  bulk.execute(writeCon, finalOptions, (err, r) => {
-    // We have connection level error
-    if (!r && err) {
-      return callback(err, null);
-    }
-
-    r.insertedCount = r.nInserted;
-    r.matchedCount = r.nMatched;
-    r.modifiedCount = r.nModified || 0;
-    r.deletedCount = r.nRemoved;
-    r.upsertedCount = r.getUpsertedIds().length;
-    r.upsertedIds = {};
-    r.insertedIds = {};
-
-    // Update the n
-    r.n = r.insertedCount;
-
-    // Inserted documents
-    const inserted = r.getInsertedIds();
-    // Map inserted ids
-    for (let i = 0; i < inserted.length; i++) {
-      r.insertedIds[inserted[i].index] = inserted[i]._id;
-    }
-
-    // Upserted documents
-    const upserted = r.getUpsertedIds();
-    // Map upserted ids
-    for (let i = 0; i < upserted.length; i++) {
-      r.upsertedIds[upserted[i].index] = upserted[i]._id;
-    }
-
-    // Return the results
-    callback(null, r);
-  });
-}
 
 // Check the update operation to ensure it has atomic operators.
 function checkForAtomicOperators(update) {
@@ -232,82 +130,6 @@ function createIndexes(coll, indexSpecs, options, callback) {
   );
 }
 
-function deleteCallback(err, r, callback) {
-  if (callback == null) return;
-  if (err && callback) return callback(err);
-  if (r == null) return callback(null, { result: { ok: 1 } });
-  r.deletedCount = r.result.n;
-  if (callback) callback(null, r);
-}
-
-/**
- * Delete multiple documents from the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter The Filter used to select the documents to remove
- * @param {object} [options] Optional settings. See Collection.prototype.deleteMany for a list of options.
- * @param {Collection~deleteWriteOpCallback} [callback] The command result callback
- */
-function deleteMany(coll, filter, options, callback) {
-  options.single = false;
-
-  removeDocuments(coll, filter, options, (err, r) => deleteCallback(err, r, callback));
-}
-
-/**
- * Delete a single document from the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter The Filter used to select the document to remove
- * @param {object} [options] Optional settings. See Collection.prototype.deleteOne for a list of options.
- * @param {Collection~deleteWriteOpCallback} [callback] The command result callback
- */
-function deleteOne(coll, filter, options, callback) {
-  options.single = true;
-  removeDocuments(coll, filter, options, (err, r) => deleteCallback(err, r, callback));
-}
-
-/**
- * Drop an index from this collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {string} indexName Name of the index to drop.
- * @param {object} [options] Optional settings. See Collection.prototype.dropIndex for a list of options.
- * @param {Collection~resultCallback} [callback] The command result callback
- */
-function dropIndex(coll, indexName, options, callback) {
-  // Delete index command
-  const cmd = { dropIndexes: coll.collectionName, index: indexName };
-
-  // Decorate command with writeConcern if supported
-  applyWriteConcern(cmd, { db: coll.s.db, collection: coll }, options);
-
-  // Execute command
-  executeCommand(coll.s.db, cmd, options, (err, result) => {
-    if (typeof callback !== 'function') return;
-    if (err) return handleCallback(callback, err, null);
-    handleCallback(callback, null, result);
-  });
-}
-
-/**
- * Drop all indexes from this collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {Object} [options] Optional settings. See Collection.prototype.dropIndexes for a list of options.
- * @param {Collection~resultCallback} [callback] The command result callback
- */
-function dropIndexes(coll, options, callback) {
-  dropIndex(coll, '*', options, err => {
-    if (err) return handleCallback(callback, err, false);
-    handleCallback(callback, null, true);
-  });
-}
-
 /**
  * Ensure that an index exists. If the index does not exist, this function creates it.
  *
@@ -319,192 +141,6 @@ function dropIndexes(coll, options, callback) {
  */
 function ensureIndex(coll, fieldOrSpec, options, callback) {
   ensureIndexDb(coll.s.db, coll.collectionName, fieldOrSpec, options, callback);
-}
-
-/**
- * Find and update a document.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} query Query object to locate the object to modify.
- * @param {array} sort If multiple docs match, choose the first one in the specified sort order as the object to manipulate.
- * @param {object} doc The fields/vals to be updated.
- * @param {object} [options] Optional settings. See Collection.prototype.findAndModify for a list of options.
- * @param {Collection~findAndModifyCallback} [callback] The command result callback
- * @deprecated use findOneAndUpdate, findOneAndReplace or findOneAndDelete instead
- */
-function findAndModify(coll, query, sort, doc, options, callback) {
-  // Create findAndModify command object
-  const queryObject = {
-    findAndModify: coll.collectionName,
-    query: query
-  };
-
-  sort = formattedOrderClause(sort);
-  if (sort) {
-    queryObject.sort = sort;
-  }
-
-  queryObject.new = options.new ? true : false;
-  queryObject.remove = options.remove ? true : false;
-  queryObject.upsert = options.upsert ? true : false;
-
-  const projection = options.projection || options.fields;
-
-  if (projection) {
-    queryObject.fields = projection;
-  }
-
-  if (options.arrayFilters) {
-    queryObject.arrayFilters = options.arrayFilters;
-    delete options.arrayFilters;
-  }
-
-  if (doc && !options.remove) {
-    queryObject.update = doc;
-  }
-
-  if (options.maxTimeMS) queryObject.maxTimeMS = options.maxTimeMS;
-
-  // Either use override on the function, or go back to default on either the collection
-  // level or db
-  options.serializeFunctions = options.serializeFunctions || coll.s.serializeFunctions;
-
-  // No check on the documents
-  options.checkKeys = false;
-
-  // Final options for retryable writes and write concern
-  let finalOptions = Object.assign({}, options);
-  finalOptions = applyRetryableWrites(finalOptions, coll.s.db);
-  finalOptions = applyWriteConcern(finalOptions, { db: coll.s.db, collection: coll }, options);
-
-  // Decorate the findAndModify command with the write Concern
-  if (finalOptions.writeConcern) {
-    queryObject.writeConcern = finalOptions.writeConcern;
-  }
-
-  // Have we specified bypassDocumentValidation
-  if (finalOptions.bypassDocumentValidation === true) {
-    queryObject.bypassDocumentValidation = finalOptions.bypassDocumentValidation;
-  }
-
-  finalOptions.readPreference = ReadPreference.primary;
-
-  // Have we specified collation
-  try {
-    decorateWithCollation(queryObject, coll, finalOptions);
-  } catch (err) {
-    return callback(err, null);
-  }
-
-  // Execute the command
-  executeCommand(coll.s.db, queryObject, finalOptions, (err, result) => {
-    if (err) return handleCallback(callback, err, null);
-
-    return handleCallback(callback, null, result);
-  });
-}
-
-/**
- * Find and remove a document.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} query Query object to locate the object to modify.
- * @param {array} sort If multiple docs match, choose the first one in the specified sort order as the object to manipulate.
- * @param {object} [options] Optional settings. See Collection.prototype.findAndRemove for a list of options.
- * @param {Collection~resultCallback} [callback] The command result callback
- * @deprecated use findOneAndDelete instead
- */
-function findAndRemove(coll, query, sort, options, callback) {
-  // Add the remove option
-  options.remove = true;
-  // Execute the callback
-  findAndModify(coll, query, sort, null, options, callback);
-}
-
-/**
- * Fetch the first document that matches the query.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} query Query for find Operation
- * @param {object} [options] Optional settings. See Collection.prototype.findOne for a list of options.
- * @param {Collection~resultCallback} [callback] The command result callback
- */
-function findOne(coll, query, options, callback) {
-  const cursor = coll
-    .find(query, options)
-    .limit(-1)
-    .batchSize(1);
-
-  // Return the item
-  cursor.next((err, item) => {
-    if (err != null) return handleCallback(callback, toError(err), null);
-    handleCallback(callback, null, item);
-  });
-}
-
-/**
- * Find a document and delete it in one atomic operation. This requires a write lock for the duration of the operation.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter Document selection filter.
- * @param {object} [options] Optional settings. See Collection.prototype.findOneAndDelete for a list of options.
- * @param {Collection~findAndModifyCallback} [callback] The collection result callback
- */
-function findOneAndDelete(coll, filter, options, callback) {
-  // Final options
-  const finalOptions = Object.assign({}, options);
-  finalOptions.fields = options.projection;
-  finalOptions.remove = true;
-  // Execute find and Modify
-  findAndModify(coll, filter, options.sort, null, finalOptions, callback);
-}
-
-/**
- * Find a document and replace it in one atomic operation. This requires a write lock for the duration of the operation.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter Document selection filter.
- * @param {object} replacement Document replacing the matching document.
- * @param {object} [options] Optional settings. See Collection.prototype.findOneAndReplace for a list of options.
- * @param {Collection~findAndModifyCallback} [callback] The collection result callback
- */
-function findOneAndReplace(coll, filter, replacement, options, callback) {
-  // Final options
-  const finalOptions = Object.assign({}, options);
-  finalOptions.fields = options.projection;
-  finalOptions.update = true;
-  finalOptions.new = options.returnOriginal !== void 0 ? !options.returnOriginal : false;
-  finalOptions.upsert = options.upsert !== void 0 ? !!options.upsert : false;
-
-  // Execute findAndModify
-  findAndModify(coll, filter, options.sort, replacement, finalOptions, callback);
-}
-
-/**
- * Find a document and update it in one atomic operation. This requires a write lock for the duration of the operation.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter Document selection filter.
- * @param {object} update Update operations to be performed on the document
- * @param {object} [options] Optional settings. See Collection.prototype.findOneAndUpdate for a list of options.
- * @param {Collection~findAndModifyCallback} [callback] The collection result callback
- */
-function findOneAndUpdate(coll, filter, update, options, callback) {
-  // Final options
-  const finalOptions = Object.assign({}, options);
-  finalOptions.fields = options.projection;
-  finalOptions.update = true;
-  finalOptions.new = typeof options.returnOriginal === 'boolean' ? !options.returnOriginal : false;
-  finalOptions.upsert = typeof options.upsert === 'boolean' ? options.upsert : false;
-
-  // Execute findAndModify
-  findAndModify(coll, filter, options.sort, update, finalOptions, callback);
 }
 
 /**
@@ -641,154 +277,6 @@ function indexInformation(coll, options, callback) {
   indexInformationDb(coll.s.db, coll.collectionName, options, callback);
 }
 
-function insertDocuments(coll, docs, options, callback) {
-  if (typeof options === 'function') (callback = options), (options = {});
-  options = options || {};
-  // Ensure we are operating on an array op docs
-  docs = Array.isArray(docs) ? docs : [docs];
-
-  // Final options for retryable writes and write concern
-  let finalOptions = Object.assign({}, options);
-  finalOptions = applyRetryableWrites(finalOptions, coll.s.db);
-  finalOptions = applyWriteConcern(finalOptions, { db: coll.s.db, collection: coll }, options);
-
-  // If keep going set unordered
-  if (finalOptions.keepGoing === true) finalOptions.ordered = false;
-  finalOptions.serializeFunctions = options.serializeFunctions || coll.s.serializeFunctions;
-
-  docs = prepareDocs(coll, docs, options);
-
-  // File inserts
-  coll.s.topology.insert(coll.s.namespace, docs, finalOptions, (err, result) => {
-    if (callback == null) return;
-    if (err) return handleCallback(callback, err);
-    if (result == null) return handleCallback(callback, null, null);
-    if (result.result.code) return handleCallback(callback, toError(result.result));
-    if (result.result.writeErrors)
-      return handleCallback(callback, toError(result.result.writeErrors[0]));
-    // Add docs to the list
-    result.ops = docs;
-    // Return the results
-    handleCallback(callback, null, result);
-  });
-}
-
-/**
- * Insert a single document into the collection. See Collection.prototype.insertOne for more information.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} doc Document to insert.
- * @param {object} [options] Optional settings. See Collection.prototype.insertOne for a list of options.
- * @param {Collection~insertOneWriteOpCallback} [callback] The command result callback
- */
-function insertOne(coll, doc, options, callback) {
-  if (Array.isArray(doc)) {
-    return callback(
-      MongoError.create({ message: 'doc parameter must be an object', driver: true })
-    );
-  }
-
-  insertDocuments(coll, [doc], options, (err, r) => {
-    if (callback == null) return;
-    if (err && callback) return callback(err);
-    // Workaround for pre 2.6 servers
-    if (r == null) return callback(null, { result: { ok: 1 } });
-    // Add values to top level to ensure crud spec compatibility
-    r.insertedCount = r.result.n;
-    r.insertedId = doc._id;
-    if (callback) callback(null, r);
-  });
-}
-
-/**
- * Inserts an array of documents into MongoDB. If documents passed in do not contain the **_id** field,
- * one will be added to each of the documents missing it by the driver, mutating the document. This behavior
- * can be overridden by setting the **forceServerObjectId** flag.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object[]} docs Documents to insert.
- * @param {number} [options] Optional settings. See Collection.prototype.insertMany for a list of options.
- * @param {Collection~insertWriteOpCallback} [callback] The command result callback
- */
-function insertMany(coll, docs, options, callback) {
-  if (!Array.isArray(docs)) {
-    return callback(
-      MongoError.create({ message: 'docs parameter must be an array of documents', driver: true })
-    );
-  }
-
-  // If keep going set unordered
-  options['serializeFunctions'] = options['serializeFunctions'] || coll.s.serializeFunctions;
-
-  docs = prepareDocs(coll, docs, options);
-
-  // Generate the bulk write operations
-  const operations = [
-    {
-      insertMany: docs
-    }
-  ];
-
-  bulkWrite(coll, operations, options, (err, result) => {
-    if (err) return callback(err, null);
-    callback(null, mapInsertManyResults(docs, result));
-  });
-}
-
-function mapInsertManyResults(docs, r) {
-  const finalResult = {
-    result: { ok: 1, n: r.insertedCount },
-    ops: docs,
-    insertedCount: r.insertedCount,
-    insertedIds: r.insertedIds
-  };
-
-  if (r.getLastOp()) {
-    finalResult.result.opTime = r.getLastOp();
-  }
-
-  return finalResult;
-}
-
-/**
- * Determine whether the collection is a capped collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {Object} [options] Optional settings. See Collection.prototype.isCapped for a list of options.
- * @param {Collection~resultCallback} [callback] The results callback
- */
-function isCapped(coll, options, callback) {
-  optionsOp(coll, options, (err, document) => {
-    if (err) return handleCallback(callback, err);
-    handleCallback(callback, null, !!(document && document.capped));
-  });
-}
-
-/**
- * Return the options of the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {Object} [options] Optional settings. See Collection.prototype.options for a list of options.
- * @param {Collection~resultCallback} [callback] The results callback
- */
-function optionsOp(coll, opts, callback) {
-  coll.s.db.listCollections({ name: coll.collectionName }, opts).toArray((err, collections) => {
-    if (err) return handleCallback(callback, err);
-    if (collections.length === 0) {
-      return handleCallback(
-        callback,
-        MongoError.create({ message: `collection ${coll.namespace} not found`, driver: true })
-      );
-    }
-
-    handleCallback(callback, err, collections[0].options || null);
-  });
-}
-
 /**
  * Return N parallel cursors for a collection to allow parallel reading of the entire collection. There are
  * no ordering guarantees for returned results.
@@ -840,175 +328,6 @@ function parallelCollectionScan(coll, options, callback) {
   });
 }
 
-// modifies documents before being inserted or updated
-function prepareDocs(coll, docs, options) {
-  const forceServerObjectId =
-    typeof options.forceServerObjectId === 'boolean'
-      ? options.forceServerObjectId
-      : coll.s.db.options.forceServerObjectId;
-
-  // no need to modify the docs if server sets the ObjectId
-  if (forceServerObjectId === true) {
-    return docs;
-  }
-
-  return docs.map(doc => {
-    if (forceServerObjectId !== true && doc._id == null) {
-      doc._id = coll.s.pkFactory.createPk();
-    }
-
-    return doc;
-  });
-}
-
-/**
- * Reindex all indexes on the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {Object} [options] Optional settings. See Collection.prototype.reIndex for a list of options.
- * @param {Collection~resultCallback} [callback] The command result callback
- */
-function reIndex(coll, options, callback) {
-  // Reindex
-  const cmd = { reIndex: coll.collectionName };
-
-  // Execute the command
-  executeCommand(coll.s.db, cmd, options, (err, result) => {
-    if (callback == null) return;
-    if (err) return handleCallback(callback, err, null);
-    handleCallback(callback, null, result.ok ? true : false);
-  });
-}
-
-function removeDocuments(coll, selector, options, callback) {
-  if (typeof options === 'function') {
-    (callback = options), (options = {});
-  } else if (typeof selector === 'function') {
-    callback = selector;
-    options = {};
-    selector = {};
-  }
-
-  // Create an empty options object if the provided one is null
-  options = options || {};
-
-  // Final options for retryable writes and write concern
-  let finalOptions = Object.assign({}, options);
-  finalOptions = applyRetryableWrites(finalOptions, coll.s.db);
-  finalOptions = applyWriteConcern(finalOptions, { db: coll.s.db, collection: coll }, options);
-
-  // If selector is null set empty
-  if (selector == null) selector = {};
-
-  // Build the op
-  const op = { q: selector, limit: 0 };
-  if (options.single) {
-    op.limit = 1;
-  } else if (finalOptions.retryWrites) {
-    finalOptions.retryWrites = false;
-  }
-
-  // Have we specified collation
-  try {
-    decorateWithCollation(finalOptions, coll, options);
-  } catch (err) {
-    return callback(err, null);
-  }
-
-  // Execute the remove
-  coll.s.topology.remove(coll.s.namespace, [op], finalOptions, (err, result) => {
-    if (callback == null) return;
-    if (err) return handleCallback(callback, err, null);
-    if (result == null) return handleCallback(callback, null, null);
-    if (result.result.code) return handleCallback(callback, toError(result.result));
-    if (result.result.writeErrors)
-      return handleCallback(callback, toError(result.result.writeErrors[0]));
-    // Return the results
-    handleCallback(callback, null, result);
-  });
-}
-
-/**
- * Rename the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {string} newName New name of of the collection.
- * @param {object} [options] Optional settings. See Collection.prototype.rename for a list of options.
- * @param {Collection~collectionResultCallback} [callback] The results callback
- */
-function rename(coll, newName, options, callback) {
-  let Collection = loadCollection();
-  // Check the collection name
-  checkCollectionName(newName);
-  // Build the command
-  const renameCollection = coll.s.namespace.toString();
-  const toCollection = coll.s.namespace.withCollection(newName);
-  const dropTarget = typeof options.dropTarget === 'boolean' ? options.dropTarget : false;
-  const cmd = { renameCollection: renameCollection, to: toCollection, dropTarget: dropTarget };
-
-  // Decorate command with writeConcern if supported
-  applyWriteConcern(cmd, { db: coll.s.db, collection: coll }, options);
-
-  // Execute against admin
-  executeDbAdminCommand(coll.s.db.admin().s.db, cmd, options, (err, doc) => {
-    if (err) return handleCallback(callback, err, null);
-    // We have an error
-    if (doc.errmsg) return handleCallback(callback, toError(doc), null);
-    try {
-      return handleCallback(
-        callback,
-        null,
-        new Collection(
-          coll.s.db,
-          coll.s.topology,
-          coll.databaseName,
-          newName,
-          coll.s.pkFactory,
-          coll.s.options
-        )
-      );
-    } catch (err) {
-      return handleCallback(callback, toError(err), null);
-    }
-  });
-}
-
-/**
- * Replace a document in the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter The Filter used to select the document to update
- * @param {object} doc The Document that replaces the matching document
- * @param {object} [options] Optional settings. See Collection.prototype.replaceOne for a list of options.
- * @param {Collection~updateWriteOpCallback} [callback] The command result callback
- */
-function replaceOne(coll, filter, doc, options, callback) {
-  // Set single document update
-  options.multi = false;
-
-  // Execute update
-  updateDocuments(coll, filter, doc, options, (err, r) => {
-    if (callback == null) return;
-    if (err && callback) return callback(err);
-    if (r == null) return callback(null, { result: { ok: 1 } });
-
-    r.modifiedCount = r.result.nModified != null ? r.result.nModified : r.result.n;
-    r.upsertedId =
-      Array.isArray(r.result.upserted) && r.result.upserted.length > 0
-        ? r.result.upserted[0] // FIXME(major): should be `r.result.upserted[0]._id`
-        : null;
-    r.upsertedCount =
-      Array.isArray(r.result.upserted) && r.result.upserted.length ? r.result.upserted.length : 0;
-    r.matchedCount =
-      Array.isArray(r.result.upserted) && r.result.upserted.length > 0 ? 0 : r.result.n;
-    r.ops = [doc];
-    if (callback) callback(null, r);
-  });
-}
-
 /**
  * Save a document.
  *
@@ -1041,143 +360,15 @@ function save(coll, doc, options, callback) {
   });
 }
 
-function updateCallback(err, r, callback) {
-  if (callback == null) return;
-  if (err) return callback(err);
-  if (r == null) return callback(null, { result: { ok: 1 } });
-  r.modifiedCount = r.result.nModified != null ? r.result.nModified : r.result.n;
-  r.upsertedId =
-    Array.isArray(r.result.upserted) && r.result.upserted.length > 0
-      ? r.result.upserted[0] // FIXME(major): should be `r.result.upserted[0]._id`
-      : null;
-  r.upsertedCount =
-    Array.isArray(r.result.upserted) && r.result.upserted.length ? r.result.upserted.length : 0;
-  r.matchedCount =
-    Array.isArray(r.result.upserted) && r.result.upserted.length > 0 ? 0 : r.result.n;
-  callback(null, r);
-}
-
-function updateDocuments(coll, selector, document, options, callback) {
-  if ('function' === typeof options) (callback = options), (options = null);
-  if (options == null) options = {};
-  if (!('function' === typeof callback)) callback = null;
-
-  // If we are not providing a selector or document throw
-  if (selector == null || typeof selector !== 'object')
-    return callback(toError('selector must be a valid JavaScript object'));
-  if (document == null || typeof document !== 'object')
-    return callback(toError('document must be a valid JavaScript object'));
-
-  // Final options for retryable writes and write concern
-  let finalOptions = Object.assign({}, options);
-  finalOptions = applyRetryableWrites(finalOptions, coll.s.db);
-  finalOptions = applyWriteConcern(finalOptions, { db: coll.s.db, collection: coll }, options);
-
-  // Do we return the actual result document
-  // Either use override on the function, or go back to default on either the collection
-  // level or db
-  finalOptions.serializeFunctions = options.serializeFunctions || coll.s.serializeFunctions;
-
-  // Execute the operation
-  const op = { q: selector, u: document };
-  op.upsert = options.upsert !== void 0 ? !!options.upsert : false;
-  op.multi = options.multi !== void 0 ? !!options.multi : false;
-
-  if (finalOptions.arrayFilters) {
-    op.arrayFilters = finalOptions.arrayFilters;
-    delete finalOptions.arrayFilters;
-  }
-
-  if (finalOptions.retryWrites && op.multi) {
-    finalOptions.retryWrites = false;
-  }
-
-  // Have we specified collation
-  try {
-    decorateWithCollation(finalOptions, coll, options);
-  } catch (err) {
-    return callback(err, null);
-  }
-
-  // Update options
-  coll.s.topology.update(coll.s.namespace, [op], finalOptions, (err, result) => {
-    if (callback == null) return;
-    if (err) return handleCallback(callback, err, null);
-    if (result == null) return handleCallback(callback, null, null);
-    if (result.result.code) return handleCallback(callback, toError(result.result));
-    if (result.result.writeErrors)
-      return handleCallback(callback, toError(result.result.writeErrors[0]));
-    // Return the results
-    handleCallback(callback, null, result);
-  });
-}
-
-/**
- * Update multiple documents in the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter The Filter used to select the documents to update
- * @param {object} update The update operations to be applied to the document
- * @param {object} [options] Optional settings. See Collection.prototype.updateMany for a list of options.
- * @param {Collection~updateWriteOpCallback} [callback] The command result callback
- */
-function updateMany(coll, filter, update, options, callback) {
-  // Set single document update
-  options.multi = true;
-  // Execute update
-  updateDocuments(coll, filter, update, options, (err, r) => updateCallback(err, r, callback));
-}
-
-/**
- * Update a single document in the collection.
- *
- * @method
- * @param {Collection} a Collection instance.
- * @param {object} filter The Filter used to select the document to update
- * @param {object} update The update operations to be applied to the document
- * @param {object} [options] Optional settings. See Collection.prototype.updateOne for a list of options.
- * @param {Collection~updateWriteOpCallback} [callback] The command result callback
- */
-function updateOne(coll, filter, update, options, callback) {
-  // Set single document update
-  options.multi = false;
-  // Execute update
-  updateDocuments(coll, filter, update, options, (err, r) => updateCallback(err, r, callback));
-}
-
 module.exports = {
-  bulkWrite,
   checkForAtomicOperators,
   createIndex,
   createIndexes,
-  deleteMany,
-  deleteOne,
-  dropIndex,
-  dropIndexes,
   ensureIndex,
-  findAndModify,
-  findAndRemove,
-  findOne,
-  findOneAndDelete,
-  findOneAndReplace,
-  findOneAndUpdate,
   group,
   indexes,
   indexExists,
   indexInformation,
-  insertMany,
-  insertOne,
-  isCapped,
-  optionsOp,
   parallelCollectionScan,
-  prepareDocs,
-  reIndex,
-  removeDocuments,
-  rename,
-  replaceOne,
-  save,
-  updateDocuments,
-  updateMany,
-  updateOne
+  save
 };
