+++ lib/operations/db_ops.js
@@ -55,21 +55,6 @@ const debugFields = [
   'noListener'
 ];
 
-// Filter out any write concern options
-const illegalCommandFields = [
-  'w',
-  'wtimeout',
-  'j',
-  'fsync',
-  'autoIndexId',
-  'strict',
-  'serializeFunctions',
-  'pkFactory',
-  'raw',
-  'readPreference',
-  'session'
-];
-
 /**
  * Add a user to the database.
  * @method
@@ -175,92 +160,6 @@ function collections(db, options, callback) {
       })
     );
   });
-}
-
-/**
- * Create a new collection on a server with the specified options. Use this to create capped collections.
- * More information about command options available at https://docs.mongodb.com/manual/reference/command/create/
- *
- * @method
- * @param {Db} db The Db instance on which to create the collection.
- * @param {string} name The collection name to create.
- * @param {object} [options] Optional settings. See Db.prototype.createCollection for a list of options.
- * @param {Db~collectionResultCallback} [callback] The results callback
- */
-function createCollection(db, name, options, callback) {
-  let Collection = loadCollection();
-
-  // Get the write concern options
-  const finalOptions = applyWriteConcern(Object.assign({}, options), { db }, options);
-
-  // Did the user destroy the topology
-  if (db.serverConfig && db.serverConfig.isDestroyed()) {
-    return callback(new MongoError('topology was destroyed'));
-  }
-
-  const listCollectionOptions = Object.assign({}, finalOptions, { nameOnly: true });
-
-  // Check if we have the name
-  db
-    .listCollections({ name }, listCollectionOptions)
-    .setReadPreference(ReadPreference.PRIMARY)
-    .toArray((err, collections) => {
-      if (err != null) return handleCallback(callback, err, null);
-      if (collections.length > 0 && finalOptions.strict) {
-        return handleCallback(
-          callback,
-          MongoError.create({
-            message: `Collection ${name} already exists. Currently in strict mode.`,
-            driver: true
-          }),
-          null
-        );
-      } else if (collections.length > 0) {
-        try {
-          return handleCallback(
-            callback,
-            null,
-            new Collection(db, db.s.topology, db.databaseName, name, db.s.pkFactory, options)
-          );
-        } catch (err) {
-          return handleCallback(callback, err);
-        }
-      }
-
-      // Create collection command
-      const cmd = { create: name };
-
-      // Decorate command with writeConcern if supported
-      applyWriteConcern(cmd, { db }, options);
-
-      // Add all optional parameters
-      for (let n in options) {
-        if (
-          options[n] != null &&
-          typeof options[n] !== 'function' &&
-          illegalCommandFields.indexOf(n) === -1
-        ) {
-          cmd[n] = options[n];
-        }
-      }
-
-      // Force a primary read Preference
-      finalOptions.readPreference = ReadPreference.PRIMARY;
-      // Execute command
-      executeCommand(db, cmd, finalOptions, err => {
-        if (err) return handleCallback(callback, err);
-
-        try {
-          return handleCallback(
-            callback,
-            null,
-            new Collection(db, db.s.topology, db.databaseName, name, db.s.pkFactory, options)
-          );
-        } catch (err) {
-          return handleCallback(callback, err);
-        }
-      });
-    });
 }
 
 /**
@@ -586,23 +485,6 @@ function indexInformation(db, name, options, callback) {
     });
 }
 
-// Transformation methods for cursor results
-function listCollectionsTransforms(databaseName) {
-  const matching = `${databaseName}.`;
-
-  return {
-    doc: doc => {
-      const index = doc.name.indexOf(matching);
-      // Remove database name if available
-      if (doc.name && index === 0) {
-        doc.name = doc.name.substr(index + matching.length);
-      }
-
-      return doc;
-    }
-  };
-}
-
 /**
  * Retrieve the current profiling information for MongoDB
  *
@@ -621,28 +503,6 @@ function profilingInfo(db, options, callback) {
   } catch (err) {
     return callback(err, null);
   }
-}
-
-/**
- * Retrieve the current profiling level for MongoDB
- *
- * @method
- * @param {Db} db The Db instance on which to retrieve the profiling level.
- * @param {Object} [options] Optional settings. See Db.prototype.profilingLevel for a list of options.
- * @param {Db~resultCallback} [callback] The command result callback
- */
-function profilingLevel(db, options, callback) {
-  executeCommand(db, { profile: -1 }, options, (err, doc) => {
-    if (err == null && doc.ok === 1) {
-      const was = doc.was;
-      if (was === 0) return callback(null, 'off');
-      if (was === 1) return callback(null, 'slow_only');
-      if (was === 2) return callback(null, 'all');
-      return callback(new Error('Error: illegal profiling level value ' + was), null);
-    } else {
-      err != null ? callback(err, null) : callback(new Error('Error with profile command'), null);
-    }
-  });
 }
 
 /**
@@ -680,40 +540,6 @@ function removeUser(db, username, options, callback) {
 
     if (err) return handleCallback(callback, err);
     handleCallback(callback, err, result);
-  });
-}
-
-/**
- * Set the current profiling level of MongoDB
- *
- * @method
- * @param {Db} db The Db instance on which to execute the command.
- * @param {string} level The new profiling level (off, slow_only, all).
- * @param {Object} [options] Optional settings. See Db.prototype.setProfilingLevel for a list of options.
- * @param {Db~resultCallback} [callback] The command result callback.
- */
-function setProfilingLevel(db, level, options, callback) {
-  const command = {};
-  let profile = 0;
-
-  if (level === 'off') {
-    profile = 0;
-  } else if (level === 'slow_only') {
-    profile = 1;
-  } else if (level === 'all') {
-    profile = 2;
-  } else {
-    return callback(new Error('Error: illegal profiling level value ' + level));
-  }
-
-  // Set up the profile number
-  command['profile'] = profile;
-
-  executeCommand(db, command, options, (err, doc) => {
-    if (err == null && doc.ok === 1) return callback(null, level);
-    return err != null
-      ? callback(err, null)
-      : callback(new Error('Error with profile command'), null);
   });
 }
 
@@ -990,7 +816,6 @@ function executeAuthRemoveUserCommand(db, username, options, callback) {
 module.exports = {
   addUser,
   collections,
-  createCollection,
   createListener,
   createIndex,
   dropCollection,
@@ -999,11 +824,8 @@ module.exports = {
   evaluate,
   executeCommand,
   executeDbAdminCommand,
-  listCollectionsTransforms,
   indexInformation,
   profilingInfo,
-  profilingLevel,
   removeUser,
-  setProfilingLevel,
   validateDatabaseName
 };
