+++ lib/operations/insert_one.js
@@ -1,12 +1,8 @@
 'use strict';
 
-const applyRetryableWrites = require('../utils').applyRetryableWrites;
-const applyWriteConcern = require('../utils').applyWriteConcern;
-const handleCallback = require('../utils').handleCallback;
 const MongoError = require('../core').MongoError;
 const OperationBase = require('./operation').OperationBase;
-const prepareDocs = require('./common_functions').prepareDocs;
-const toError = require('../utils').toError;
+const insertDocuments = require('./common_functions').insertDocuments;
 
 class InsertOneOperation extends OperationBase {
   constructor(collection, doc, options) {
@@ -38,38 +34,6 @@ class InsertOneOperation extends OperationBase {
       if (callback) callback(null, r);
     });
   }
-}
-
-function insertDocuments(coll, docs, options, callback) {
-  if (typeof options === 'function') (callback = options), (options = {});
-  options = options || {};
-  // Ensure we are operating on an array op docs
-  docs = Array.isArray(docs) ? docs : [docs];
-
-  // Final options for retryable writes and write concern
-  let finalOptions = Object.assign({}, options);
-  finalOptions = applyRetryableWrites(finalOptions, coll.s.db);
-  finalOptions = applyWriteConcern(finalOptions, { db: coll.s.db, collection: coll }, options);
-
-  // If keep going set unordered
-  if (finalOptions.keepGoing === true) finalOptions.ordered = false;
-  finalOptions.serializeFunctions = options.serializeFunctions || coll.s.serializeFunctions;
-
-  docs = prepareDocs(coll, docs, options);
-
-  // File inserts
-  coll.s.topology.insert(coll.s.namespace, docs, finalOptions, (err, result) => {
-    if (callback == null) return;
-    if (err) return handleCallback(callback, err);
-    if (result == null) return handleCallback(callback, null, null);
-    if (result.result.code) return handleCallback(callback, toError(result.result));
-    if (result.result.writeErrors)
-      return handleCallback(callback, toError(result.result.writeErrors[0]));
-    // Add docs to the list
-    result.ops = docs;
-    // Return the results
-    handleCallback(callback, null, result);
-  });
 }
 
 module.exports = InsertOneOperation;
