+++ lib/operations/list_collections.js
@@ -4,10 +4,25 @@ const CommandOperationV2 = require('./command_v2');
 const Aspect = require('./operation').Aspect;
 const defineAspects = require('./operation').defineAspects;
 const maxWireVersion = require('../core/utils').maxWireVersion;
-const listCollectionsTransforms = require('./db_ops').listCollectionsTransforms;
 const CONSTANTS = require('../constants');
 
 const LIST_COLLECTIONS_WIRE_VERSION = 3;
+
+function listCollectionsTransforms(databaseName) {
+  const matching = `${databaseName}.`;
+
+  return {
+    doc: doc => {
+      const index = doc.name.indexOf(matching);
+      // Remove database name if available
+      if (doc.name && index === 0) {
+        doc.name = doc.name.substr(index + matching.length);
+      }
+
+      return doc;
+    }
+  };
+}
 
 class ListCollectionsOperation extends CommandOperationV2 {
   constructor(db, filter, options) {
